+++ SwiftCompilerSources/Sources/Optimizer/InstructionSimplification/SimplifyCondFail.swift
@@ -20,54 +20,18 @@ extension CondFailInst : OnoneSimplifiable, SILCombineSimplifiable {
       return
     }
 
+    // Eliminates
+    // ```
+    //   %0 = integer_literal 0
+    //   cond_fail %0, "message"
+    // ```
     guard let literal = condition as? IntegerLiteralInst,
-          let value = literal.value
+          let value = literal.value,
+          value == 0
     else {
       return
     }
 
-    if value == 0 {
-      // Eliminates
-      // ```
-      //   %0 = integer_literal 0
-      //   cond_fail %0, "message"
-      // ```
-      context.erase(instruction: self)
-    } else {
-
-      // Inserts an unreachable after an unconditional fail:
-      // ```
-      //   %0 = integer_literal 1
-      //   cond_fail %0, "message"
-      //   // following instructions
-      // ```
-      // ->
-      // ```
-      //   %0 = integer_literal 1
-      //   cond_fail %0, "message"
-      //   unreachable
-      // deadblock:
-      //   // following instructions
-      // ```
-      if InstructionList(first: self.next).allSatisfy({ $0.isUnreachableOrEndingLifetime }) {
-        // Don't remove instructions which would be re-inserted by lifetime completion.
-        return
-      }
-      let builder = Builder(after: self, context)
-      let unreachable = builder.createUnreachable()
-      _ = context.splitBlock(after: unreachable)
-    }
-  }
-}
-
-private extension Instruction {
-  var isUnreachableOrEndingLifetime: Bool {
-    switch self {
-    case is EndBorrowInst, is DestroyValueInst, is EndLifetimeInst, is DeallocStackInst, is EndAccessInst,
-         is UnreachableInst:
-      return true
-    default:
-      return false
-    }
+    context.erase(instruction: self)
   }
 }
