+++ include/swift/SILOptimizer/PassManager/Passes.def
@@ -79,6 +79,8 @@ PASS(ComputeEscapeEffects, "compute-escape-effects",
      "Computes function escape effects")
 PASS(ComputeSideEffects, "compute-side-effects",
      "Computes function side effects")
+PASS(CondFailOptimization, "cond-fail-optimization",
+     "Optimizes cond_fail instructions")
 PASS(DiagnoseInfiniteRecursion, "diagnose-infinite-recursion",
      "Diagnose Infinitely-Recursive Code")
 PASS(InitializeStaticGlobals, "initialize-static-globals",
