+++ lib/SILOptimizer/PassManager/PassPipeline.cpp
@@ -384,6 +384,7 @@ void addSimplifyCFGSILCombinePasses(SILPassPipelinePlan &P) {
   // Jump threading can expose opportunity for silcombine (enum -> is_enum_tag->
   // cond_br).
   P.addSILCombine();
+  P.addCondFailOptimization();
   // Which can expose opportunity for simplifycfg.
   P.addSimplifyCFG();
 }
