+++ test/SILOptimizer/sil_combine_ossa.sil
@@ -2243,11 +2243,10 @@ bb0(%0 : $Int):
   return %8 : $()
 }
 
-// CHECK-LABEL: @remove_pointer_compare_to_zero_NE :
-// CHECK-NOT: apply
-// CHECK: cond_fail
-// CHECK: unreachable
-// CHECK: } // end sil function 'remove_pointer_compare_to_zero_NE'
+// CHECK-LABEL: sil [ossa] @remove_pointer_compare_to_zero_NE
+// CHECK:         %1 = integer_literal $Builtin.Int1, -1
+// CHECK:         cond_fail %1
+// CHECK:       } // end sil function 'remove_pointer_compare_to_zero_NE'
 sil [ossa] @remove_pointer_compare_to_zero_NE : $@convention(thin) (Int) -> () {
 bb0(%0 : $Int):
   %1 = string_literal utf8 "ss"
@@ -2895,22 +2894,6 @@ sil public_external [serialized] @materializeForSetClosure : $@convention(thin) 
 bb0(%0 : $Builtin.RawPointer, %1 : $*Builtin.UnsafeValueBuffer, %2 : $*GenContainer<T>, %3 : $@thick GenContainer<T>.Type):
   unreachable
 }
-
-
-// CHECK-LABEL:   sil [ossa] @remove_dead_code_after_cond_fail :
-// CHECK:           [[Ref:%.*]] = integer_literal $Builtin.Int1, -1
-// CHECK-NEXT:      cond_fail [[Ref]]
-// CHECK-NEXT:      unreachable
-// CHECK: } // end sil function 'remove_dead_code_after_cond_fail'
-sil [ossa] @remove_dead_code_after_cond_fail : $@convention(thin) () -> (Int32) {
-bb0:
-  %0 = integer_literal $Builtin.Int32, -2
-  %1 = integer_literal $Builtin.Int1, -1
-  cond_fail %1 : $Builtin.Int1
-  %3 = struct $Int32 (%0 : $Builtin.Int32)
-  return %3 : $Int32
-}
-
 
 // CHECK-LABEL:   sil [ossa] @dont_remove_code_after_cond_fail
 // CHECK:         bb0([[Cond:%.*]] : $Builtin.Int1):
@@ -5679,5 +5662,77 @@ bb0(%0 : @owned $S1):
   %5 = struct $MoveOnlyStruct (%2), forwarding: @owned
   destroy_value [dead_end] %5
   unreachable
+}
+
+struct MyArray<T> {
+  var storage: MyArrayStorage
+}
+
+class MyArrayStorage {
+}
+
+struct MySliceBuffer<T> {
+  var owner: AnyObject
+  var subscriptBaseAddress: UnsafeMutablePointer<T>
+}
+
+struct MyArraySlice<T> {
+  var buffer: MySliceBuffer<T>
+}
+
+// Just check that SILCombine doesn't produce illegal SIL.
+sil [ossa] @dont_cut_off_control_flow_after_cond_fail : $@convention(thin) @async (@in_guaranteed MyArray<S1>) -> () {
+bb0(%0 : $*MyArray<S1>):
+  %2 = load_borrow %0
+  %3 = struct_extract %2, #MyArray.storage
+  %7 = copy_value %3
+  %8 = init_existential_ref %7 : $MyArrayStorage : $MyArrayStorage, $AnyObject
+  %9 = copy_value %8
+  %10 = struct $MySliceBuffer<S1> (%9, undef : $UnsafeMutablePointer<S1>)
+  end_borrow %2
+  %12 = struct $MyArraySlice<S1> (%10)
+  %13 = copy_value %12
+  %14 = destructure_struct %13
+  %15 = begin_borrow [lexical] %14
+  cond_br undef, bb2, bb1
+
+bb1:
+  end_borrow %15
+  destroy_value %14
+  br bb3
+
+bb2:
+  %20 = struct_extract %15, #MySliceBuffer.owner
+  %21 = copy_value %20
+  destroy_value %21
+  end_borrow %15
+  destroy_value %14
+  br bb3
+
+bb3:
+  %26 = begin_borrow [lexical] %12
+  cond_br undef, bb4, bb5
+
+bb4:
+  %28 = integer_literal $Builtin.Int1, -1
+  cond_fail %28, "Division results in an overflow"
+  end_borrow %26
+  destroy_value [dead_end] %12
+  destroy_value [dead_end] %8
+  unreachable
+
+bb5:
+  %34 = struct_extract %26, #MyArraySlice.buffer
+  %35 = struct_extract %34, #MySliceBuffer.subscriptBaseAddress
+  %36 = struct_extract %34, #MySliceBuffer.owner
+  %37 = mark_dependence %35 on %36
+  %38 = struct_extract %37, #UnsafeMutablePointer._rawValue
+  %39 = metatype $@thin S1.Type
+  %40 = builtin "copyArray"<S1>(%39, undef : $Builtin.RawPointer, %38, undef : $Builtin.Word) : $()
+  destroy_value %8
+  end_borrow %26
+  destroy_value %12
+  %44 = tuple ()
+  return %44
 }
