+++ test/SILOptimizer/sil_combine_uncheck_ossa.sil
@@ -27,9 +27,9 @@ bb0(%0 : $Builtin.Int1):
 }
 
 // CHECK-LABEL: sil [ossa] @test_unconditional_checked_cast_addr_fail
-// CHECK: bb0
-// CHECK: unreachable
-// CHECK: }
+// CHECK:         cond_fail
+// CHECK-NEXT:    unconditional_checked_cast_addr
+// CHECK:       } // end sil function 'test_unconditional_checked_cast_addr_fail'
 sil [ossa] @test_unconditional_checked_cast_addr_fail : $@convention(thin) (@in E, @in D) -> @out D {
 bb0(%0 : $*D, %1 : $*E, %2 : $*D):
   unconditional_checked_cast_addr E in %1 : $*E to D in %0 : $*D
